pyproject.toml
[project]
name = "escrapper-backend"
version = "0.1.0"
description = "E-Paper News Aggregation & Visualization System — FastAPI Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    # Database (async)
    "sqlalchemy[asyncio]>=2.0.28",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    # Data Validation
    "pydantic>=2.6.3",
    "pydantic-settings>=2.2.1",
    # Task Queue
    "celery>=5.3.6",
    "redis>=5.0.3",
    # Web Scraping
    "playwright>=1.43.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",
    # NLP / Processing
    "nltk>=3.8.1",
    "scikit-learn>=1.4.1",
    # HTTP Utilities
    "httpx>=0.27.0",
    "tenacity>=8.2.3",
    # Logging
    "structlog>=24.1.0",
    # CORS / Security
    "python-dotenv>=1.0.1",
    "pdfplumber>=0.11.9",
    "fpdf2>=2.7.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "ruff>=0.3.4",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["../tests/backend"]

[tool.setuptools.packages.find]
where = ["backend"]
namespaces = true

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
